requirements.txt
pywebio==1.3.3
psycopg2-binary==2.9.1
pandas==1.3.1
azure-identity==1.4.0
python-dotenv==0.15.0
azure-storage-file-datalake==12.3.0
requests==2.25.1
requests-oauthlib==1.3.0
azure-identity==1.4.0
pytest==6.2.4
emojis==0.6.0
progress==1.5
numpy
pandas-profiling==3.0.0
ipywidgets==7.6.3
sweetviz==2.1.3
apache-airflow==2.1.3
apache-airflow-providers-ftp==2.0.0
apache-airflow-providers-http==2.0.0
apache-airflow-providers-imap==2.0.0
apache-airflow-providers-sqlite==2.0.0
notebook==6.4.3
